+++ app/layouts/default.vue
@@ -251,13 +251,14 @@ async function getEngine(uaParser = parser) {
 }
 
 function setWindowStore() {
-    store.isMobile = window.innerWidth < 700;
+    store.isMobile = window.innerWidth < 700 || window.innerHeight < 500;
     store.isMobileOrTablet = window.innerWidth < 1365;
-    store.isTablet = window.innerWidth < 1365 && window.innerWidth >= 700;
+    store.isTablet = window.innerWidth < 1365 && window.innerWidth >= 700 && window.innerHeight > 500;
     store.isPC = window.innerWidth >= 1365;
     store.isPCWide = window.innerWidth >= 1900;
     store.scrollbarWidth = window.innerWidth - document.documentElement.offsetWidth;
     store.viewport.width = window.innerWidth;
+    store.viewport.height = window.innerHeight;
 }
 
 const listener = () => {